feat: remove CommandModule and add various AuthService implementations
Removed the CommandModule and updated the AuthModule to use a new AuthService. Replaced SyncRolesPermissionCommand with an AuthService implemented in AuthorizeSupabaseService.ts. Also, new methods for token creation, decoding, authorization, and user authentication are added in IAuthorizeService.ts. Installation and use of the jwt-simple package for handling JWTs introduced as part of these modifications. Additionally, PermissionInterceptor.ts and IDecodeToken.ts were added to handle permission-based requests and JWT token decoding respectively.